\name{getLabel}
\alias{getLabel}

\title{
Get label of an entity based on the URI.
}
\description{
Get label of an entity base on the URI
}
\usage{
getLabel(uri, endpoint = "http://www.ebi.ac.uk/rdf/services/atlas/sparql")
}

\arguments{
  \item{uri}{
URI of the class for which the label is required.
}
  \item{endpoint}{
        The location of the SPARQL endpoint - default value is live Atlas endpoint
  }

}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\author{
James Malone, Simon Jupp
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

}

