\name{getTaxonURI}
\alias{getTaxonURI}

\title{
Get URI of a taxon class based on the common or taxonomic name.
}
\description{
Convenience method to the ontology URI, used in other queries, based on an input common or taxonomic name such as "human" or "Homo sapiens"
}
\usage{
getTaxonURI(taxonName)
}

\arguments{
  \item{taxonName}{
Common or taxonomic name
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\author{
James Malone, Simon Jupp
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

}

