\name{transcription_pathway_enrichment}
\alias{transcription_pathway_enrichment}
\docType{data}

\title{Expression Gene Set Enrichment result for }
\description{
A result set example from running the Atlas Expression gene set enrichment function. The gene list was originally extracted form a transcription pathways in Reactome and enrichment performed to see which experimental factors were enriched for this pathway.
}
\usage{transcription_pathway_enrichment}
\format{A list of enrichmentresult objects}
\source{Gene Expression Atlas RDF at European Bioinformatics Institute}
\keyword{datasets}